requirements.txt
# Core runtime
gradio>=4.44.0
ultralytics>=8.3.0
opencv-python>=4.9.0.80
pillow>=10.3.0
numpy>=1.26.4
# On CPU, Ultralytics pulls a compatible Torch. If you need a specific CUDA build, install torch separately.
# For local Windows webcam builds, you can swap to:
# opencv-python>=4.9.0.80